# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AdReceipt Agent API",
    "version": "2.0.0",
    "description": "Create policy-bound sponsored placements and verify their settlement receipts. All paths are served through the application's same-origin /api proxy. Placement preparation never broadcasts a transaction."
  },
  "servers": [{ "url": "/api" }],
  "paths": {
...
    "/receipts": {
      "get": {
        "operationId": "listVerifiedReceipts",
        "summary": "List live receipts corroborated by The Graph and Sepolia RPC",
        "responses": {
          "200": { "description": "Verified receipt ledger and Graph health" },
          "503": { "description": "Ledger evidence unavailable" }
        }
      }
    },
    "/receipts/{receiptId}": {
      "get": {
        "operationId": "verifyReceipt",
        "summary": "Verify one receipt by its EIP-712 digest",
        "parameters": [{ "$ref": "#/components/parameters/ReceiptId" }],
        "responses": {
          "200": { "description": "PAID_VERIFIED, pending, not found, or invalid result" },
          "503": { "description": "Graph or RPC evidence unavailable" }
        }
      }
    },
    "/subjects/{subjectHash}": {
      "get": {
        "operationId": "verifyRecommendationSubject",
        "summary": "Find and verify receipts for an exact recommendation commitment",
        "parameters": [
          {
            "name": "subjectHash",
            "in": "path",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Bytes32" }
          }
        ],
        "responses": {
          "200": { "description": "Verified receipts and deterministic disclosure decision" },
          "503": { "description": "Graph or RPC evidence unavailable" }
        }
      }
    }
  },
  "components": {
    "parameters": {
...
      "ReceiptId": {
        "name": "receiptId",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/Bytes32" }
      }
    },
    "schemas": {
      "Bytes32": { "type": "string", "pattern": "^0x[0-9a-fA-F]{64}$" },
...
    }
  }
}